Fixed floating point return value for argument matching scoring algo.

+++ src/main/java/com/cedarsoftware/util/io/MetaUtils.java
@@ -1045,7 +1045,7 @@ public class MetaUtils
             return 1;
         }
 
-        return found / parameters.length;
+        return (double) found / parameters.length;
     }
 
     private static boolean setParameterIfPossible(Object[] arguments, List<ParameterHint> list, int i, String name) {
